Fix reek
