+++ lib/tramway/helpers/decorate_helper.rb
@@ -8,22 +8,6 @@ module Tramway
       def tramway_decorate(object_or_array, decorator: nil)
         decorator_class(object_or_array, decorator).decorate object_or_array, self
       end
-
-      private
-
-      def decorator_class(object_or_array, decorator)
-        if decorator.present?
-          decorator
-        else
-          klass = if Tramway::Decorators::CollectionDecorators.collection?(object_or_array)
-                    object_or_array.first.class
-                  else
-                    object_or_array.class
-                  end
-
-          "#{klass}Decorator".constantize
-        end
-      end
     end
   end
 end
