+++ spec/dummy/app/decorators/user_decorator.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
+# Test decorator User model as a default
 class UserDecorator < Tramway::BaseDecorator
 end
