+++ spec/dummy/app/decorators/user_specific_decorator.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
+# Test specific decorator for User model
 class UserSpecificDecorator < Tramway::BaseDecorator
 end
