+++ spec/dummy/app/models/user.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
+# Test model
 class User < ApplicationRecord
 end
